rename stakingHook file
